requirements.txt
pymongo
langchain_openai
Python-IO
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
python-dotenv
regex
APScheduler
apify-client
